ARTCC zones and tiers
Airspace in the United States is broken up into several zones. These are
called Air Route Traffic Control Centers (ARTCCs), with each zone having
first and second "tiers" of zones they are associated with in their
vicinity.

Having this information will be important for generating friendly
messages since programs, ground stops and delays since these messages
are often scoped to the zone within which the airport resides.

Example: In the test file that I have in this repository, there is
currently a ground stop applied to EWR with a scope applying to 1st tier
airports surrounding it. Subsequently, this means that even though EWR
has a ground stop applied to it, flights taking off from other airports
in the NYC metro (LGA, JFK, DCA, etc.) will also be stopped. This is
important information to show in a friendly message!